Move file_transfer_token validation from core to common module
