+++ concent_api/gatekeeper/views.py
@@ -18,12 +18,12 @@ from golem_messages.message         import Message
 from golem_messages.shortcuts       import load
 
 from core.exceptions import FileTransferTokenError
-from core.validation import validate_file_transfer_token
 from gatekeeper.utils               import gatekeeper_access_denied_response
 from common                          import logging
 from common.constants                import ErrorCode
 from common.decorators import provides_concent_feature
 from common.helpers import get_current_utc_timestamp
+from common.validations import validate_file_transfer_token
 
 
 logger = getLogger(__name__)
